fix zfs build mv failure
if the build is interrupted and re-ran, the build.sh script tries to move the spl-{version} directory (not just it's contents) inside of directory spl, instead of renaming it to spl, so you get spl/spl-{version}. this stops the build in it's tracks no matter how many times it is restarted. this commit makes it so the contents of spl-{version} are simply moved on top of spl if it already exists, or renames spl-{version} to spl as intended if not.